Directly merge a patch from Martin Vahlensieck:
This puts the N flag in the synopsis of the man page.  Also order the
flags with the upper case option first (in the list and in the programs
usage) like it is done in other manual pages such as ls(1).

While here fix a trailing whitespace mandoc -Tlint found.