Accept standalone leading asterisks.
Add a hack around SQLite 3.41.2 (and possibly others) that have a single
leading asterisk instead of two asterisks.  This is...  probably?... a
typo.

This also clarifies the warning message for when there's a single
asterisk without subsequent characters.

It also refactors the check for this behaviour into a single function.

Fixes #4