Render <table> as TS/TE.
This causes instances of <table> to render as TS/TE statements,
with rows and columns rendered appropriately as well.  Tables
within <blockquote>, which seem to be the norm, are simply output
as-is; while tables without the context are bracketed by sp
vertical space.

While here, apply a small fix to the order of HTML tag array
elements.